# api-sys_produtos/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Produtos-api Documentation",
        "description": "api-sys_produtos",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8989",
            "description": "local server"
        }
    ],
    "paths": {
...
        "/api/produto/get/{id}": {
            "get": {
                "tags": [
                    "Produto"
                ],
                "summary": "Get Produto",
                "description": "Produtos Controller",
                "operationId": "1219e6c9ca2b743fecafbed1a55ba866",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/produto/create": {
            "post": {
                "tags": [
                    "Produto"
                ],
                "summary": "Create Produto",
                "description": "Produtos Controller",
                "operationId": "aa896566cccf1f2df094671c4bd36422",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "double"
                                    },
                                    "date_expiration": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "image": {
                                        "type": "string"
                                    },
                                    "id_categoria": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/produto/update": {
            "post": {
                "tags": [
                    "Produto"
                ],
                "summary": "Update Produto",
                "description": "Produtos Controller",
                "operationId": "bb5b2ca62635f52e0181aa37eb1b773c",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "double"
                                    },
                                    "date_expiration": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "image": {
                                        "type": "string"
                                    },
                                    "id_categoria": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}